pyproject.toml
[project]
name = "mas-sycophancy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "gdm-concordia",
    "google-cloud-aiplatform>=1.60.0",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "pandas",
    "matplotlib",
    "seaborn",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]